+++ src/main/SqlExpression/MySQL/Enum/Placeholder.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace WebArch\BitrixOrmTools\SqlExpression\MySQL\Enum;
+
+class Placeholder
+{
+    const STRING = '?s';
+
+    const INTEGER = '?i';
+
+    const FLOAT = '?f';
+
+    const COLUMN = '?#';
+}